+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- 口座のtypeOfをAccountで固定
+- PermitのtypeOfをPermitで固定
+
 ### Deprecated
 
 ### Removed